+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release notes
 
-## v3.3.0 (2026-03-03)
+## v3.3.0 (2026-02-28)
 
 - ✨ Added `hasValue()` global function which checks if variables are defined and have non-nullable values. If variable inside is "" empty string, 0 int, 0.0 float, nil, {} empty object, [] empty array or false, it returns `false`. Closes [#73](https://github.com/textwire/textwire/issues/73).
 - ✨ Added `format()` built-in function for strings that allows you to embed values into a string. `{{ "He has %s apples".format(apples.len()) }}`. Closes [#71](https://github.com/textwire/textwire/issues/71).